### 2024-11-27 #7535 @editorialbot
[PRE REVIEW]: `CausalTables.jl`: Simulating and storing data for statistical causal inference in Julia
**Submitting author:** @salbalkus (<a href="http://orcid.org/0000-0003-4695-833X">Salvador Balkus</a>)
**Repository:** https://github.com/salbalkus/CausalTables.jl
**Branch with paper.md** (empty if default branch): main
**Version:** v1.1.1
