### 2024-11-27 #7535 @crvernon
:wave: @salbalkus - Please remove any formatting off the the title of your paper.  This causes some issues in how our automated system handles things.  Thanks.
