### 2024-11-27 97a8a13
remove formatting from paper title
