### 2024-12-08 #7535 @salbalkus
@lrnv Absolutely. It's unfortunate that the editorialbot detected all R packages; I think @blind-contours would be a good reviewer for the topic if they know Julia.

@olivierlabayle may be interested in reviewing since this paper cites TMLE.jl.

Otherwise, looking through the JOSS database, some potential reviewers include @juliohm, @epiben, and @gdalle. Would any of you be interested in reviewing this submission?
