### 2024-12-08 #7535 @juliohm
I can find some time to review until the end of the year. Would that be ok?

Em dom., 8 de dez. de 2024, 16:55, Salvador Balkus ***@***.***>
escreveu:

> @lrnv <https://github.com/lrnv> Absolutely. It's unfortunate that the
> editorialbot detected all R packages; I think @blind-contours
> <https://github.com/blind-contours> would be a good reviewer for the
> topic if they know Julia.
>
> @olivierlabayle <https://github.com/olivierlabayle> may be interested in
> reviewing since this paper cites TMLE.jl.
>
> Otherwise, looking through the JOSS database, some potential reviewers
> include @juliohm <https://github.com/juliohm>, @epiben
> <https://github.com/epiben>, and @gdalle <https://github.com/gdalle>.
> Would any of you be interested in reviewing this submission?
>
