### 2024-12-11 #7580 @editorialbot
[REVIEW]: `CausalTables.jl`: Simulating and storing data for statistical causal inference in Julia
**Submitting author:** @salbalkus (<a href="http://orcid.org/0000-0003-4695-833X">Salvador Balkus</a>)
**Repository:** https://github.com/salbalkus/CausalTables.jl
**Branch with paper.md** (empty if default branch): main
**Version:** 1.2.2
