### 2024-12-23 #7580 @juliohm
@salbalkus thank you for the package!

Please consider the following review comments.

## Reproducibility

The examples in the paper are not fully reproducible:

```julia
julia> ate(scm) # average treatment effect
(μ = 1.000, eff_bound = 2.000)
```

versus

![image](https://github.com/user-attachments/assets/3cc8c46e-8972-4ae6-9996-6480a7767569)

I ran the exact same preamble:

```julia
using Printf
Base.show(io::IO, f::Float64) = @printf(io, "%.3f", f)
using Random
Random.seed!(1)
```

Could this be caused by a difference in the Julia version or execution platform?

```
Julia Version 1.11.2
Commit 5e9a32e7af (2024-12-01 20:02 UTC)
Build Info:
  Official https://julialang.org/ release
Platform Info:
  OS: Windows (x86_64-w64-mingw32)
  CPU: 8 × Intel(R) Core(TM) i5-1035G4 CPU @ 1.10GHz
  WORD_SIZE: 64
  LLVM: libLLVM-16.0.6 (ORCJIT, icelake-client)
Threads: 1 default, 0 interactive, 1 GC (on 8 virtual cores)
```
