### 2024-12-23 #7580 @juliohm
## Functionality

The README states that

```
CausalTables.jl is a tool for to evaluating and compare the statistical performance of causal inference methods in Julia.
```

It is not clear how the package helps with the evaluation and comparison of causal **inference methods**. Could you please elaborate on this functionality claim? The examples in the documentation illustrate the representation and simulation of **causal models**, plus statistical summaries and ground truth estimates.
