### 2024-12-23 #7580 @juliohm
I left some items unchecked in my review checklist (https://github.com/openjournals/joss-reviews/issues/7580#issuecomment-2558715821), which depend on further work on your side @salbalkus.
