### 2024-12-23 #7580 @olivierlabayle
# State of the field

The paper describes some limitations of other packages but does not name them so it is difficult to verify these, e.g. "packages have inconsitent API". Are there any other Julia/Python package that have similar goals?
