### 2024-12-23 #7580 @olivierlabayle
# Functionality

What types of structural causal models are supported by CausalTables.jl? The paper and documentation presents the classic (A, W, Y) structure. It seems that multiple treatments, with or without their own set of confounders are not suported.

```julia
ERROR: ArgumentError: More than one treatment not allowed
```

I have not been able to add additional outcome covariates in the CausalTable either. This is limiting the integration of the present package with TMLE.jl for instance, which natively supports these [causal structures](https://targene.github.io/TMLE.jl/stable/user_guide/scm/) and [estimands](https://targene.github.io/TMLE.jl/stable/user_guide/estimands/).

A brief look at the code seems to indicate that the package is strongly based on the `treatment`, `response`, `confounders` paradigm (no DAG) so that I am not sure how this would extend well to more sophisticated causal structures. I think this scope should be made explicit.

Also, if this is the case, I think the following should be updated to better reflect current possibilities, limitations and opportunities

"The flexibility of CausalTables.jl allows users to easily extract ground truth values for any relevant aspect of a data-generating process. This supports benchmarking causal estimators of virtually any estimand that fits in the SCM framework, not just those mentioned in these two examples.".
