### 2025-01-07 f2154cc
update paper and README
