+++ paper/paper.md
@@ -12,7 +12,7 @@ authors:
   name: Nima S. Hejazi
   orcid: 0000-0002-7127-2789
 bibliography: paper.bib
-date: 2024-11-26
+date: 2025-01-07
 tags:
 - Julia
 - statistics
@@ -26,18 +26,19 @@ toc-title: Table of contents
 # Summary
 
 Estimating the strength of causal relationships between variables is an
-important problem across many scientific disciplines. A variety of
-statistical methods have been developed to estimate and obtain
-inferences about causal quantities, yet few tools readily support the
-comparison of candidate approaches. `CausalTables.jl` offers tools to
-evaluate and compare statistical causal inference methods in Julia. The
-package provides two main functionalities. Firstly, it implements a
-`CausalTable` interface for storing data with partially-labeled causal
-structure in a `Tables.jl`-compatible format. Secondly, it introduces a
-`StructuralCausalModel` for randomly generating data with a
-user-specified causal structure and computing ground truth parameters
-under the given experiment. Together, these functionalities expand the
-Julia ecosystem by supporting the use and benchmarking of the growing
+important problem across many scientific disciplines. `CausalTables.jl`
+supports the development of new statistical methods for causal inference
+in Julia by providing tools to (1) easily store and process data endowed
+with causal structure and (2) simulate data from causal models for
+experimental testing. Firstly, the package implements a `CausalTable`
+structure that stores features annotated with causal labels in a
+`Tables.jl`-compatible format. Its interface includes causal-relevant
+functions, such as extracting relevant variables and applying
+interventions on treatment. Secondly, `CausalTables.jl` introduces a
+`StructuralCausalModel` for randomly generating data from user-specified
+causal models and computing ground truth parameters under the given
+experiment. Together, these functionalities expand the Julia ecosystem
+by supporting the development and experimental assessment of the growing
 number of causal inference methods.
 
 # Statement of need
@@ -48,53 +49,67 @@ relationships between variables from observed data
 [@pearl2009causality; @hernan2020causal]; causal inference techniques
 have helped applied scientists and decision-makers better understand
 important phenomena in fields ranging from health and medicine to
-politics and economics. As interest in causal inference has grown across
-many disciplines, so too has the development of software tools for
-estimating causal effects. While Julia packages for causal inference
-have begun to emerge---with examples including, for estimation,
-`TMLE.jl` [@TMLE.jl] and `CausalELM.jl` [@CausalELM.jl], and, for causal
-discovery, `CausalInference.jl` [@Schauer2024]---the ecosystem is still
-in its infancy. New methods for causal inference are being developed at
-a rapid pace, underscoring the need for tools designed to support the
-evaluation and comparison of their performance. `CausalTables.jl` aims
-to provide such a tool for the Julia language. Currently, attempts to
-benchmark causal inference methods in Julia face two major challenges.
+politics and economics. As interest in causal inference continues to
+grow across many disciplines, so too does the development of software
+tools for estimating causal effects. While Julia packages for causal
+inference have begun to emerge---with examples including `TMLE.jl`
+[@TMLE.jl] and `CausalELM.jl`[@CausalELM.jl] for estimation and
+`CausalInference.jl` [@Schauer2024] for causal discovery---the ecosystem
+is still in its infancy. New methods for causal inference are being
+developed at a rapid pace, underscoring the need for tools designed to
+support their development. `CausalTables.jl` aims to provide such a tool
+for the Julia language. Currently, attempts to implement and test causal
+inference methods in Julia face two major challenges.
 
[…]
