### 2025-01-07 #7580 @salbalkus
Hi @juliohm,

Thank you for your review! I updated the paper and README to hopefully address your comments.

Regarding **[Reproducibility](https://github.com/openjournals/joss-reviews/issues/7580#issuecomment-2558750526)**, the `ate` function (along with most of the other example functions) is a Monte Carlo procedure that does involve some randomness, so I believe this is due to the Julia version. Upon investigating, it seems that my Quarto installation that I used to generate the .md file for the paper was running an older version of Julia. Once I updated it to use the latest version, it appears to give the same result for the particular code block you posted. Please let me know if there are any other blocks of code giving inconsistent results as well, and I can check it out.

Regarding **[Functionality](https://github.com/openjournals/joss-reviews/issues/7580#issuecomment-2558771791)**, the main use of the package is storing and simulating data assumed to be drawn from causal models (what you described), but the hope is that these functions will make it easier for statisticians to implement and benchmark causal inference methods, i.e. by making it easier to process the data and write simulations. I've updated the README to more clearly reflect this.

All causal inference methods must assume some underlying causal model, hence why this package is intended for the broader field of "causal inference" and not just "causal models", but you are correct that the package isn't explicitly comparing causal inference methods. I've also updated the paper to try and more clearly reflect this idea, and bring the "storing data and simulating from SCMs (structural causal models) in the field of causal inference" purpose to the forefront. Let me know if this still isn't clear enough, or there are any other issues, and I can resolve them.
