### 2025-01-07 b409a1b
emphasize tabular data in paper
