+++ paper/paper.md
@@ -28,18 +28,18 @@ toc-title: Table of contents
 Estimating the strength of causal relationships between variables is an
 important problem across many scientific disciplines. `CausalTables.jl`
 supports the development of new statistical methods for causal inference
-in Julia by providing tools to (1) easily store and process data endowed
-with causal structure and (2) simulate data from causal models for
-experimental testing. Firstly, the package implements a `CausalTable`
-structure that stores features annotated with causal labels in a
-`Tables.jl`-compatible format. Its interface includes causal-relevant
-functions, such as extracting relevant variables and applying
-interventions on treatment. Secondly, `CausalTables.jl` introduces a
-`StructuralCausalModel` for randomly generating data from user-specified
-causal models and computing ground truth parameters under the given
-experiment. Together, these functionalities expand the Julia ecosystem
-by supporting the development and experimental assessment of the growing
-number of causal inference methods.
+in Julia by providing tools to (1) easily store and process tabular data
+endowed with causal structure and (2) simulate data from causal models
+for experimental testing. Firstly, the package implements a
+`CausalTable` structure that stores features annotated with causal
+labels in a `Tables.jl`-compatible format. Its interface includes
+causal-relevant functions, such as extracting relevant variables and
+applying interventions on treatment. Secondly, `CausalTables.jl`
+introduces a `StructuralCausalModel` for randomly generating data from
+user-specified causal models and computing ground truth parameters under
+the given experiment. Together, these functionalities expand the Julia
+ecosystem by supporting the development and experimental assessment of
+the growing number of causal inference methods.
 
 # Statement of need
 
@@ -56,10 +56,11 @@ inference have begun to emerge---with examples including `TMLE.jl`
 [@TMLE.jl] and `CausalELM.jl`[@CausalELM.jl] for estimation and
 `CausalInference.jl` [@Schauer2024] for causal discovery---the ecosystem
 is still in its infancy. New methods for causal inference are being
-developed at a rapid pace, underscoring the need for tools designed to
-support their development. `CausalTables.jl` aims to provide such a tool
-for the Julia language. Currently, attempts to implement and test causal
-inference methods in Julia face two major challenges.
+developed at a rapid pace, but there currently do not exist auxiliary
+tools designed to support their development in the Julia language.
+`CausalTables.jl` aims to provide such a tool. Currently, attempts to
+implement and test causal inference methods in Julia face two major
+challenges.
 
 First, causal inference requires data to be preprocessed in various ways
 based on the underlying causal structure. Suppose one were to write
@@ -92,11 +93,11 @@ of several common causal effect parameters.
 
 By addressing these two major challenges---preprocessing and
 simulation--- `CausalTables.jl` simplifies and accelerates the
-development of tools for statistical causal inference in Julia. The
-`CausalTable` interface extends `Tables.jl`, the most common interface
-for accessing tabular data in Julia [@quinn2024tables]. The SCM
-framework operates in conjunction with `Distributions.jl`, the primary
-Julia package for working with random variables
+development of tools for statistical causal inference on tabular data in
+Julia. The `CausalTable` interface extends `Tables.jl`, the most common
+interface for accessing tabular data in Julia [@quinn2024tables]. The
+SCM framework operates in conjunction with `Distributions.jl`, the
+primary Julia package for working with random variables
 [@JSSv098i16; @Distributions.jl-2019]. By integrating seamlessly with
 other commonly used packages in the Julia ecosystem, `CausalTables.jl`
 ensures both compatibility and ease of use for statisticians and applied
@@ -338,19 +339,19 @@ mean(conmean(scm, ct_intervened, :Y) .- responsematrix(ct))
 # Closing remarks
 
 `CausalTables.jl` provides useful auxiliary functions to support causal
[…]
