+++ paper/paper.qmd
@@ -24,17 +24,17 @@ engine: julia
 
 # Summary
 
-Estimating the strength of causal relationships between variables is an important problem across many scientific disciplines. `CausalTables.jl` supports the development of new statistical methods for causal inference in Julia by providing tools to (1) easily store and process data endowed with causal structure and (2) simulate data from causal models for experimental testing. Firstly, the package implements a `CausalTable` structure that stores features annotated with causal labels in a `Tables.jl`-compatible format. Its interface includes causal-relevant functions, such as extracting relevant variables and applying interventions on treatment. Secondly, `CausalTables.jl` introduces a `StructuralCausalModel` for randomly generating data from user-specified causal models and computing ground truth parameters under the given experiment. Together, these functionalities expand the Julia ecosystem by supporting the development and experimental assessment of the growing number of causal inference methods. 
+Estimating the strength of causal relationships between variables is an important problem across many scientific disciplines. `CausalTables.jl` supports the development of new statistical methods for causal inference in Julia by providing tools to (1) easily store and process tabular data endowed with causal structure and (2) simulate data from causal models for experimental testing. Firstly, the package implements a `CausalTable` structure that stores features annotated with causal labels in a `Tables.jl`-compatible format. Its interface includes causal-relevant functions, such as extracting relevant variables and applying interventions on treatment. Secondly, `CausalTables.jl` introduces a `StructuralCausalModel` for randomly generating data from user-specified causal models and computing ground truth parameters under the given experiment. Together, these functionalities expand the Julia ecosystem by supporting the development and experimental assessment of the growing number of causal inference methods. 
 
 # Statement of need
 
-The quantitative science of causal inference has emerged over the past three decades as a set of formalisms for studying cause-and-effect relationships between variables from observed data [@pearl2009causality; @hernan2020causal]; causal inference techniques have helped applied scientists and decision-makers better understand important phenomena in fields ranging from health and medicine to politics and economics. As interest in causal inference continues to grow across many disciplines, so too does the development of software tools for estimating causal effects. While Julia packages for causal inference have begun to emerge---with examples including `TMLE.jl` [@TMLE.jl] and `CausalELM.jl`[@CausalELM.jl] for estimation and `CausalInference.jl` [@Schauer2024] for causal discovery---the ecosystem is still in its infancy. New methods for causal inference are being developed at a rapid pace, underscoring the need for tools designed to support their development. `CausalTables.jl` aims to provide such a tool for the Julia language. Currently, attempts to implement and test causal inference methods in Julia face two major challenges.
+The quantitative science of causal inference has emerged over the past three decades as a set of formalisms for studying cause-and-effect relationships between variables from observed data [@pearl2009causality; @hernan2020causal]; causal inference techniques have helped applied scientists and decision-makers better understand important phenomena in fields ranging from health and medicine to politics and economics. As interest in causal inference continues to grow across many disciplines, so too does the development of software tools for estimating causal effects. While Julia packages for causal inference have begun to emerge---with examples including `TMLE.jl` [@TMLE.jl] and `CausalELM.jl`[@CausalELM.jl] for estimation and `CausalInference.jl` [@Schauer2024] for causal discovery---the ecosystem is still in its infancy. New methods for causal inference are being developed at a rapid pace, but there currently do not exist auxiliary tools designed to support their development in the Julia language. `CausalTables.jl` aims to provide such a tool. Currently, attempts to implement and test causal inference methods in Julia face two major challenges.
 
[…]
