### 2025-01-07 9f484c7
add comparison to existing packages
