+++ paper/paper.bib
@@ -148,3 +148,44 @@
   copyright = {Creative Commons Attribution 4.0 International}
 }
 
+@manual{squires2018causaldag,
+  title={{\texttt{causaldag}: creation, manipulation, and learning of causal models}},
+  author={{Chandler Squires}},
+  year={2018},
+  url={https://github.com/uhlerlab/causaldag},
+}
+
+ @article{Textor2017, 
+  title={Robust causal inference using directed acyclic graphs: the R package ‘dagitty’}, 
+  ISSN={1464-3685}, url={http://dx.doi.org/10.1093/ije/dyw341}, 
+  DOI={10.1093/ije/dyw341}, journal={International Journal of Epidemiology}, 
+  publisher={Oxford University Press (OUP)}, 
+  author={Textor, Johannes and van der Zander, Benito and Gilthorpe, Mark S. and Liśkiewicz, Maciej and Ellison, George T.H.}, 
+  year={2017}, 
+  month=jan, 
+  pages={dyw341} }
+
+@book{tlverse,
+  title={Targeted Learning in R: Causal Data Science with the tlverse Software Ecosystem},
+  author={Mark {van der Laan} and Jeremy Coyle and Nima Hejazi and Ivana Malenica and Rachael Phillips and Alan Hubbard},
+  year={2009},
+  publisher={Cambridge university press},
+  url={https://tlverse.org/tlverse-handbook/}
+}
+
+@article{Chen2020,
+  title        = {{CausalML}: Python Package for Causal Machine Learning},
+  author       = {Chen, Huigang and Harinen, Totte and Lee, Jeong-Yoon and
+                  Yung, Mike and Zhao, Zhenyu},
+  year         =  {2020},
+  journal       = {arXiv preprint arXiv:2002.11631}
+}
+
+@article{dowhy,
+  title={DoWhy: An End-to-End Library for Causal Inference},
+  author={Sharma, Amit and Kiciman, Emre},
+  journal={arXiv preprint arXiv:2011.04216},
+  year={2020}
+}
+
+
