+++ paper/paper.qmd
@@ -28,13 +28,17 @@ Estimating the strength of causal relationships between variables is an importan
 
 # Statement of need
 
-The quantitative science of causal inference has emerged over the past three decades as a set of formalisms for studying cause-and-effect relationships between variables from observed data [@pearl2009causality; @hernan2020causal]; causal inference techniques have helped applied scientists and decision-makers better understand important phenomena in fields ranging from health and medicine to politics and economics. As interest in causal inference continues to grow across many disciplines, so too does the development of software tools for estimating causal effects. While Julia packages for causal inference have begun to emerge---with examples including `TMLE.jl` [@TMLE.jl] and `CausalELM.jl`[@CausalELM.jl] for estimation and `CausalInference.jl` [@Schauer2024] for causal discovery---the ecosystem is still in its infancy. New methods for causal inference are being developed at a rapid pace, but there currently do not exist auxiliary tools designed to support their development in the Julia language. `CausalTables.jl` aims to provide such a tool. Currently, attempts to implement and test causal inference methods in Julia face two major challenges.
+The quantitative science of causal inference has emerged over the past three decades as a set of formalisms for studying cause-and-effect relationships between variables from observed data [@pearl2009causality; @hernan2020causal]; causal inference techniques have helped applied scientists and decision-makers better understand important phenomena in fields ranging from health and medicine to politics and economics. New methods for causal inference are being developed at a rapid pace, but there currently do not exist auxiliary tools designed to support their development in the Julia language. `CausalTables.jl` aims to provide such a tool. Presently, attempts to implement and test causal inference methods in Julia face two major challenges.
 
 First, causal inference requires data to be preprocessed in various ways based on the underlying causal structure. Suppose one were to write their own method by building on existing statistical packages in Julia. Using `MLJ.jl` [@blaom2020mlj] would necessitate extracting the treatment and response as Vectors and the variables hypothesized to cause them as Tables; meanwhile; using `GLM.jl` [@bates2023glm] would require the same but as Matrix or `Formula` objects. This challenge has also led to differences in the API for existing causal methods: for instance, `CausalELM.jl` [@CausalELM.jl] requires the user to split apart treatment and response variables as individual vectors, while `TMLE.jl` require the entire dataset in a `Tables.jl`-compliant format with treatment and response variables being labeled via strings or symbols. `CausalTables.jl` provides a `CausalTable` interface that, by packaging the data and auxiliary causal knowledge together, allows extracting relevant causal components in multiple ways. This simplifies both writing new packages as well as processing data to evaluate existing packages.
 
 Second, testing the performance of new estimators often requires simulating data for numerical experiments from a Structural Causal Model (SCM) [@pearl2009causality] so as to compare them to an underlying ground truth (encoded via interventions on the SCM). An SCM defines causal structure by envisaging a data-generating process as random draws from a sequence of non-parametric structural equations, with each draw depending on realizations from draws preceding it. `CausalTables.jl` provides a simple, user-friendly way to define an SCM, sample data randomly from it, and compute or approximate the underlying true values of several common causal effect parameters.
 
[…]
