### 2025-01-07 #7580 @salbalkus
Hi @olivierlabayle,

Thank you for your review! Some responses to your comments:

Regarding [**State of the field**](https://github.com/openjournals/joss-reviews/issues/7580#issuecomment-2559300339), it seems the paper was initially not clear about the limitations of existing work --- "packages have inconsistent API" was meant more as a comment on the challenge of implementing and comparing packages for causal inference, unrelated to the state of the field. I've clarified what I mean by this by modifying the Statement of Need section of the paper.

In addition, from what I've seen, Julia packages for causal inference tend to implement their own subset of SCM functionality; hence, I am not aware of any other packages for simulating and processing tabular data with causal structure. The most similar existing packages focus on specifying a causal DAG (`DoWhy` and `CausalDAG` in Python, `daggity` in R, `CausalInference` in Julia) but these are quite different in that they work on graphs; they do not seem to support simulating data nor performing common data manipulation tasks on tabular data for causal inference. Either way, to address your comment I have created a new section "Comparison to existing packages" that clarifies the state of the field more clearly. Let me know if this is adequate or you feel it needs expansion.

Regarding **[Functionality](https://github.com/openjournals/joss-reviews/issues/7580#issuecomment-2559358365)**, there are several points that I will comment on:
- The inability to create causal tables with multiple treatments and outcomes is unintended behavior; the package should support multiple, so this will be fixed soon.
- The package will assume, at a baseline, that the causal inference problem involves at least one treatment and response; I've tried to clarify this more with an update to the "Closing Remarks" section.
- That said, this package is intended to go beyond just "one-treatment-response-confounders" paradigm. Reading the TMLE.jl documentation, I have realized that the inability to specify multiple treatments with *different* sets of confounders is something that has been overlooked in CausalTables.jl, and I plan to implement this in response to your comment.
- To support the above, along with labels for other types of causal variables (mediators, instruments, etc.), it seems necessary to tweak the current underlying representation of the causal structure to allow at least a partial DAG-like structure. Do you mind if I draw some inspiration from the setup used by `TMLE.jl` [that you linked previously](https://targene.github.io/TMLE.jl/stable/user_guide/scm/)? Specifically, I think it would be efficient to adopt something under the hood similar to data structure being passed as input in this example:

```
SCM([
	:Y => [:T₁, :T₂, :W₁₁, :W₁₂, :W₂₁, :W₂₂, :W, :C],
	:T₁ => [:W₁₁, :W₁₂, :W],
	:T₂ => [:W₂₁, :W₂₂, :W] ])
```

I would like this package to be reasonably interoperable with `TMLE.jl`, and storing the underlying causal structure in this way seems like it would make it easy to integrate them. Just let me know, and I will get back to you once I've updated the package to support the types of causal structures needed for TMLE.jl.
