+++ paper/paper.bib
@@ -150,7 +150,7 @@
 
 @manual{squires2018causaldag,
   title={{\texttt{causaldag}: creation, manipulation, and learning of causal models}},
-  author={{Chandler Squires}},
+  author={Chandler Squires},
   year={2018},
   url={https://github.com/uhlerlab/causaldag},
 }
@@ -168,7 +168,7 @@
 @book{tlverse,
   title={Targeted Learning in R: Causal Data Science with the tlverse Software Ecosystem},
   author={Mark {van der Laan} and Jeremy Coyle and Nima Hejazi and Ivana Malenica and Rachael Phillips and Alan Hubbard},
-  year={2009},
+  year={2024},
   publisher={Cambridge university press},
   url={https://tlverse.org/tlverse-handbook/}
 }
