+++ paper/paper.qmd
@@ -38,7 +38,7 @@ By addressing these two major challenges---preprocessing and simulation--- `Caus
 
 # Comparison to existing packages
 
-As interest in causal inference continues to grow across disciplines, so too has the development of software tools for estimating causal effects. While the a multitude of methods have been implemented in the R and Python languages (for instance, [@tlverse] or [@Chen2020]), Julia has seen relatively fewer. Recent Julia packages for causal inference include `TMLE.jl` [@TMLE.jl] and `CausalELM.jl`[@CausalELM.jl]. These packages focus on estimation techniques using tabular data: they implement specific ways to label causal structure, but do not provide a general simulation or causal-specific data processing interface like `CausalTables.jl`. On the other hand, `CausalInference.jl` [@Schauer2024] provides an interface for representing causal graphs and implements causal discovery algorithms, similar to CausalDAG [@squires2018causaldag] or DoWhy [@dowhy] in Python and daggity [@Textor2017] in R. However, it is generally incompatible with the tabular data format required by statistical tools, and also cannot simulate data. In fact, as far as we are aware, `CausalTables.jl` is the first package for simulating and extracting ground-truth causal estimands from an existing SCM in Julia.
+As interest in causal inference continues to grow across disciplines, so too has the development of software tools for estimating causal effects. While the a multitude of methods have been implemented in the R and Python languages [@tlverse; @Chen2020], Julia has seen relatively fewer. Recent Julia packages for causal inference include `TMLE.jl` [@TMLE.jl] and `CausalELM.jl` [@CausalELM.jl]. These packages focus on estimation techniques using tabular data: they implement specific ways to label causal structure, but do not provide a general simulation or causal-specific data processing interface like `CausalTables.jl`. On the other hand, `CausalInference.jl` [@Schauer2024] provides an interface for representing causal graphs and implements causal discovery algorithms, similar to CausalDAG [@squires2018causaldag] or DoWhy [@dowhy] in Python and daggity [@Textor2017] in R. However, it is generally incompatible with the tabular data format required by statistical tools, and also cannot simulate data. In fact, as far as we are aware, `CausalTables.jl` is the first package for simulating and extracting ground-truth causal estimands from an existing SCM in Julia.
 
 # Instructional use cases
 
