### 2025-01-08 #7580 @lrnv
@salbalkus I see htat the wordcount of the paper is about 2000, which is twice as long as recomended. Since we are still early in the review, I have to ask you to reduce the lenght by half. Joss papers are not documentation nor full-lenght exposition of research, please re-read [the docs](https://joss.readthedocs.io/en/latest/paper.html). Maybe the extra material can be salvaged into extra documentation for your package instead of thrown out completely ?
