+++ paper/paper.md
@@ -12,7 +12,7 @@ authors:
   name: Nima S. Hejazi
   orcid: 0000-0002-7127-2789
 bibliography: paper.bib
-date: 07 January 2025
+date: 08 January 2025
 tags:
 - Julia
 - statistics
@@ -27,71 +27,48 @@ toc-title: Table of contents
 
 Estimating the strength of causal relationships between variables is an
 important problem across many scientific disciplines. `CausalTables.jl`
-supports the development of new statistical methods for causal inference
-in Julia by providing tools to (1) easily store and process tabular data
-endowed with causal structure and (2) simulate data from causal models
-for experimental testing. Firstly, the package implements a
-`CausalTable` structure that stores features annotated with causal
-labels in a `Tables.jl`-compatible format. Its interface includes
-causal-relevant functions, such as extracting relevant variables and
-applying interventions on treatment. Secondly, `CausalTables.jl`
-introduces a `StructuralCausalModel` for randomly generating data from
-user-specified causal models and computing ground truth parameters under
-the given experiment. Together, these functionalities expand the Julia
-ecosystem by supporting the development and experimental assessment of
-the growing number of causal inference methods.
+provides tools to (1) easily store and process tabular data endowed with
+causal structure and (2) simulate data from causal models for
+experimental testing and compute ground-truth estimates. Together, these
+functionalities expand the Julia ecosystem by supporting the development
+and experimental assessment of new statistical causal inference methods
+in Julia.
 
 # Statement of need
 
 The quantitative science of causal inference has emerged over the past
 three decades as a set of formalisms for studying cause-and-effect
 relationships between variables from observed data
-[@pearl2009causality; @hernan2020causal]; causal inference techniques
+[@pearl2009causality; @hernan2020causal]. Causal inference techniques
 have helped applied scientists and decision-makers better understand
 important phenomena in fields ranging from health and medicine to
-politics and economics. New methods for causal inference are being
-developed at a rapid pace, but there currently do not exist auxiliary
-tools designed to support their development in the Julia language.
-`CausalTables.jl` aims to provide such a tool. Presently, attempts to
-implement and test causal inference methods in Julia face two major
-challenges.
+politics and economics. New software tools for causal inference are
+being developed at a rapid pace, but in the Julia language, there
+currently do not exist auxiliary tools designed to support their
+development. `CausalTables.jl` aims to provide such a tool.
 
-First, causal inference requires data to be preprocessed in various ways
-based on the underlying causal structure. Suppose one were to write
-their own method by building on existing statistical packages in Julia.
-Using `MLJ.jl` [@blaom2020mlj] would necessitate extracting the
-treatment and response as Vectors and the variables hypothesized to
-cause them as Tables; meanwhile; using `GLM.jl` [@bates2023glm] would
-require the same but as Matrix or `Formula` objects. This challenge has
-also led to differences in the API for existing causal methods: for
-instance, `CausalELM.jl` [@CausalELM.jl] requires the user to split
-apart treatment and response variables as individual vectors, while
-`TMLE.jl` require the entire dataset in a `Tables.jl`-compliant format
-with treatment and response variables being labeled via strings or
-symbols. `CausalTables.jl` provides a `CausalTable` interface that, by
-packaging the data and auxiliary causal knowledge together, allows
-extracting relevant causal components in multiple ways. This simplifies
[…]
