+++ paper/paper.qmd
@@ -16,7 +16,7 @@ authors:
 affiliations:
  - name: Department of Biostatistics, Harvard T.H. Chan School of Public Health
    index: 1
-date: 07 January 2025
+date: 08 January 2025
 bibliography: paper.bib
 format: markdown
 engine: julia
@@ -24,27 +24,23 @@ engine: julia
 
 # Summary
 
-Estimating the strength of causal relationships between variables is an important problem across many scientific disciplines. `CausalTables.jl` supports the development of new statistical methods for causal inference in Julia by providing tools to (1) easily store and process tabular data endowed with causal structure and (2) simulate data from causal models for experimental testing. Firstly, the package implements a `CausalTable` structure that stores features annotated with causal labels in a `Tables.jl`-compatible format. Its interface includes causal-relevant functions, such as extracting relevant variables and applying interventions on treatment. Secondly, `CausalTables.jl` introduces a `StructuralCausalModel` for randomly generating data from user-specified causal models and computing ground truth parameters under the given experiment. Together, these functionalities expand the Julia ecosystem by supporting the development and experimental assessment of the growing number of causal inference methods. 
+Estimating the strength of causal relationships between variables is an important problem across many scientific disciplines. `CausalTables.jl` provides tools to (1) easily store and process tabular data endowed with causal structure and (2) simulate data from causal models for experimental testing and compute ground-truth estimates. Together, these functionalities expand the Julia ecosystem by supporting the development and experimental assessment of new statistical causal inference methods in Julia. 
 
 # Statement of need
 
-The quantitative science of causal inference has emerged over the past three decades as a set of formalisms for studying cause-and-effect relationships between variables from observed data [@pearl2009causality; @hernan2020causal]; causal inference techniques have helped applied scientists and decision-makers better understand important phenomena in fields ranging from health and medicine to politics and economics. New methods for causal inference are being developed at a rapid pace, but there currently do not exist auxiliary tools designed to support their development in the Julia language. `CausalTables.jl` aims to provide such a tool. Presently, attempts to implement and test causal inference methods in Julia face two major challenges.
+The quantitative science of causal inference has emerged over the past three decades as a set of formalisms for studying cause-and-effect relationships between variables from observed data [@pearl2009causality; @hernan2020causal]. Causal inference techniques have helped applied scientists and decision-makers better understand important phenomena in fields ranging from health and medicine to politics and economics. New software tools for causal inference are being developed at a rapid pace, but in the Julia language, there currently do not exist auxiliary tools designed to support their development. `CausalTables.jl` aims to provide such a tool. 
 
[…]
