+++ paper/paper.bib
@@ -166,10 +166,9 @@
   pages={dyw341} }
 
 @book{tlverse,
-  title={Targeted Learning in R: Causal Data Science with the tlverse Software Ecosystem},
+  title={{Targeted Learning in R}},
   author={Mark {van der Laan} and Jeremy Coyle and Nima Hejazi and Ivana Malenica and Rachael Phillips and Alan Hubbard},
   year={2024},
-  publisher={Cambridge university press},
   url={https://tlverse.org/tlverse-handbook/}
 }
 
