### 2025-01-08 #7580 @salbalkus
@lrnv I cut a lot of redundant and overly-technical content from the paper and removed one of the SCM examples. Do you think it is still too long?

In an ideal world I'd like to include *some* code or examples because I think it's a little hard to understand what the package actually does without a demonstration, but if you still think the paper is too long I can remove these examples (the documentation already has similar demonstrations).
