### 2025-01-09 1afb36c
remove "confounders", replace with "causes" attribute
