+++ test/runtests.jl
@@ -24,7 +24,8 @@ end
     foo3 = Tables.rowtable(foo1)
 
     # DataFrame form
-    df = CausalTables.CausalTable(foo1, :X, :Y)    
+    df = CausalTables.CausalTable(foo1, :X, :Y)  
+    
     @test Tables.istable(df)
     @test Tables.columntable(foo1) == df.data
     @test ncol(df) == 3
@@ -34,7 +35,7 @@ end
     @test Tables.getcolumn(df, 1) == X
     @test Tables.columnindex(df, :X) == 1
     @test Tables.columntype(df, :X) == Int
-    @test df.confounders == [:Z]
+    @test df.causes == (X = [:Z], Y = [:Z, :X])
 
     @test Tables.columnaccess(df)
     @test Tables.rowaccess(df)
@@ -42,7 +43,7 @@ end
     @test Tables.schema(df) == Tables.schema(data(df))
 
     # Row table form
-    rowtbl = CausalTables.CausalTable(foo3, :X, :Y; confounders = [:Z])
+    rowtbl = CausalTables.CausalTable(foo3, :X, :Y; causes = (X = [:Z], Y = [:X, :Z]))
     @test Tables.istable(rowtbl)
     @test Tables.rowtable(rowtbl.data) == foo3
     @test ncol(rowtbl) == 3
@@ -53,7 +54,7 @@ end
     @test Tables.columntype(rowtbl, :X) == Int
 
     # Column table form
-    coltbl = CausalTables.CausalTable(foo2, :X, :Y; confounders = [:Z])
+    coltbl = CausalTables.CausalTable(foo2, :X, :Y, (X = [:Z], Y = [:X, :Z]))
     @test Tables.istable(coltbl)
     @test coltbl.data == foo2
     @test ncol(coltbl) == 3
@@ -65,17 +66,14 @@ end
     @test Tables.columnnames(coltbl) == (:X, :Y, :Z)
 
     # Extra causal-related functions
-    coltbl.response
     more_sums = (S = Sum(:X, :G), T = Sum(:Z, :G), U = Sum(:Y, :G))
     coltbl2 = CausalTables.replace(coltbl, arrays = (G = [1 0 1; 0 1 1; 0 0 1],), summaries = more_sums)
-    coltbl2  = summarize(coltbl2) 
-
+    coltbl2 = summarize(coltbl2) 
     @test coltbl2.treatment == [:X, :S]
-    @test coltbl2.confounders == [:Z, :T]
+    @test coltbl2.causes == (X = [:Z, :T], Y = [:X, :Z, :S, :T], S = [:Z, :T], U = [:X, :Z, :S, :T])
     @test coltbl2.response == [:Y, :U]
     @test Tables.columnnames(CausalTables.treatment(coltbl2)) == (:X, :S)
     @test Tables.columnnames(CausalTables.response(coltbl2)) == (:Y, :U)
-    @test Tables.columnnames(CausalTables.confounders(coltbl2)) == (:Z, :T)
     @test Tables.columnnames(CausalTables.treatmentparents(coltbl2)) == (:Z, :T)
     @test Tables.columnnames(CausalTables.parents(coltbl2, :X)) == (:Z, :T)
     @test Tables.columnnames(CausalTables.parents(coltbl2, :Z)) == ()
@@ -85,9 +83,10 @@ end
     # Other convenience
     baz = (X = [4, 5], Y = ["foo", "bar"], Z = [0.1, 0.2])
     array = Graphs.adjacency_matrix(Graphs.path_graph(2))
-    causalbaz = CausalTables.CausalTable(baz, :X, :Y, [:Z], (A = array,), (B = CausalTables.Sum(:X, :A),))
[…]
