+++ src/causal_table.jl
@@ -1,85 +1,112 @@
-function _process_causal_variable_names(treatment, response, confounders)
-    ## Process treatment and response variables into vectors, if they are not already vectors
-    if typeof(treatment) <: Symbol
+function _process_causal_variable_names(treatment, response, causes)
+
+    # Wrap symbols in lists if they haven't been already
+    if treatment isa Symbol
         treatment = [treatment]
     end
 
-    if typeof(response) <: Symbol
+    if response isa Symbol
         response = [response]
     end
 
-    ## Error Handling ##
-
-    # Assume only univariate treatment and response
-    # TODO: Future code should allow for multivariate treatment and response
-    # TODO: Undo temporary disabled line of code by checking for either 1 treatment/response or 2 treatments/response with 1 summarized
-    #length(treatment) > 1 && throw(ArgumentError("Only univariate treatment is currently supported"))
-    #length(response) > 1 && throw(ArgumentError("Only univariate response is currently supported"))
-
-    # Ensure treatment, response, and confounders do not overlap
-    name_occurrences = StatsBase.countmap(vcat(treatment, response, confounders))
+    # Ensure treatment and response do not overlap
+    name_occurrences = StatsBase.countmap(vcat(treatment, response))
     name_repeats = filter(((k, v),) -> v > 1, name_occurrences)
-    length(name_repeats) > 0 && throw(ArgumentError("The following variable names are repeated across treatment, response, and confounder labels: $(keys(name_repeats))")) 
+    length(name_repeats) > 0 && throw(ArgumentError("The following variable names are repeated across treatment and response lists: $(keys(name_repeats))")) 
     
+    if !isnothing(causes)
+        # Check that `causes` is acyclic
+        _check_dag(causes) && throw(ArgumentError("`causes` contains a cycle, but causal relationships must form a directed acyclic graph (DAG), meaning no cycles are allowed."))
+
+        # Ensure response does not cause any variables in the DAG
+        response_is_a_cause = any.(map(r -> r .∈ values(causes), response))
+        any(response_is_a_cause) && throw(ArgumentError("`causes` denotes a response variables causing a treatment variable, which is not allowed."))
+    end
+
     # Return fully processed causal variable names
-    return treatment, response, confounders
+    return treatment, response, causes
+end
+
+# Function that outputs true if the input is not a directed acyclic graph
+function _check_dag(causes_original)
+    causes = deepcopy(causes_original)
+
+    S = setdiff(keys(causes), vcat(values(causes)...))
+    L = Set{Symbol}()
+
+    while(length(S) > 0)
+        s = pop!(S)
+        push!(L, s)
+        if s ∈ keys(causes)
+            while length(causes[s]) > 0
+                e = pop!(causes[s])
+                if(e ∈ keys(causes))
+                    push!(S, e)
+                end
+            end
+        end
+    end
+    return any(length.(values(causes)) .!= 0)
+end
+
+# Default assumption: if not specified, set anything not labeled as a cause of all treatments and all responses
+function set_unlabeled_causes(data, treatment, response)
+    u = vcat(treatment, response)
+    confounders = setdiff(Tables.columnnames(data), u)
+    causes = Dict()
+
+    for t in treatment
+        causes[t] = confounders
+    end
+
+    confounders_and_treatment = vcat(confounders, treatment)
+    for r in response
+        causes[r] = confounders_and_treatment
+    end
+    return (;causes...) # Unpack dictionary into NamedTuple
 end
 
 mutable struct CausalTable
-
     # data storage
     data::NamedTuple
 
     # labels
     treatment::Symbols
     response::Symbols
[…]
