### 2025-01-09 310c2df
update references to "confounders", including removal from documentation
