+++ docs/src/man/formatting.md
@@ -15,7 +15,7 @@ using DataFrames
 tbl = BostonHousing().dataframe
 
 # Wrapping the dataset in a CausalTable
-ctbl = CausalTable(tbl; treatment = :NOX, response = :MEDV, confounders = [:CRIM, :ZN, :INDUS, :CHAS, :B, :DIS, :LSTAT])
+ctbl = CausalTable(tbl; treatment = :NOX, response = :MEDV)
 
 nothing # hide
 ```
@@ -38,7 +38,7 @@ responseparents(ctbl)        # get CausalTable of treatment and confounders
 data(ctbl)                   # get underlying wrapped dataset
 
 # replace one or more attributes of the CausalTable
-CausalTables.replace(ctbl; response = :CRIM, confounders = [:MEDV, :ZN, :INDUS, :CHAS, :B, :DIS, :LSTAT]) 
+CausalTables.replace(ctbl; response = :CRIM) 
 
 nothing # hide
 ```
