### 2025-01-09 44f8a5f
allow select to return an empty CausalTable if passed empty Array; require cause labels for treatment and response
