+++ docs/src/man/formatting.md
@@ -37,8 +37,10 @@ confounders(ctbl)            # get CausalTable of confounders
 responseparents(ctbl)        # get CausalTable of treatment and confounders
 data(ctbl)                   # get underlying wrapped dataset
 
-# replace one or more attributes of the CausalTable
-CausalTables.replace(ctbl; response = :CRIM) 
+# Replace one or more attributes of the CausalTable.
+# Setting `causes = nothing` is a quick shortcut to specify
+# that all unlabeled variables are confounders of the treatment-response relationship
+CausalTables.replace(ctbl; response = :CRIM, causes = nothing) 
 
 nothing # hide
 ```
