### 2025-01-09 badeb7c
update JOSS badge
