### 2025-01-09 8a03fe7
implement mediators and lay groundwork for instruments
