+++ test/runtests.jl
@@ -24,8 +24,7 @@ end
     foo3 = Tables.rowtable(foo1)
 
     # DataFrame form
-    df = CausalTables.CausalTable(foo1, :X, :Y)  
-    
+    df = CausalTables.CausalTable(foo1, [:X, :Z], :Y, (X = [:Z], Z = [], Y = [:X, :Z]))  
     @test Tables.istable(df)
     @test Tables.columntable(foo1) == df.data
     @test ncol(df) == 3
@@ -35,7 +34,7 @@ end
     @test Tables.getcolumn(df, 1) == X
     @test Tables.columnindex(df, :X) == 1
     @test Tables.columntype(df, :X) == Int
-    @test df.causes == (X = [:Z], Y = [:Z, :X])
+    @test df.causes == (X = [:Z], Z = [], Y = [:X, :Z])
 
     @test Tables.columnaccess(df)
     @test Tables.rowaccess(df)
@@ -43,7 +42,7 @@ end
     @test Tables.schema(df) == Tables.schema(data(df))
 
     # Row table form
-    rowtbl = CausalTables.CausalTable(foo3, :X, :Y; causes = (X = [:Z], Y = [:X, :Z]))
+    rowtbl = CausalTables.CausalTable(foo3, :X, [:Y, :Z]; causes = (X = [], Y = [:X, :Z], Z = [:X]))
     @test Tables.istable(rowtbl)
     @test Tables.rowtable(rowtbl.data) == foo3
     @test ncol(rowtbl) == 3
@@ -54,7 +53,7 @@ end
     @test Tables.columntype(rowtbl, :X) == Int
 
     # Column table form
-    coltbl = CausalTables.CausalTable(foo2, :X, :Y, (X = [:Z], Y = [:X, :Z]))
+    coltbl = CausalTables.CausalTable(foo2, :X, :Y)
     @test Tables.istable(coltbl)
     @test coltbl.data == foo2
     @test ncol(coltbl) == 3
@@ -70,14 +69,14 @@ end
     coltbl2 = CausalTables.replace(coltbl, arrays = (G = [1 0 1; 0 1 1; 0 0 1],), summaries = more_sums)
     coltbl2 = summarize(coltbl2) 
     @test coltbl2.treatment == [:X, :S]
-    @test coltbl2.causes == (X = [:Z, :T], Y = [:X, :Z, :S, :T], S = [:Z, :T], U = [:X, :Z, :S, :T])
+    @test coltbl2.causes == (X = [:Z, :T], Y = [:Z, :X, :T, :S], S = [:Z, :T], U = [:Z, :X, :T, :S])
     @test coltbl2.response == [:Y, :U]
     @test Tables.columnnames(CausalTables.treatment(coltbl2)) == (:X, :S)
     @test Tables.columnnames(CausalTables.response(coltbl2)) == (:Y, :U)
     @test Tables.columnnames(CausalTables.treatmentparents(coltbl2)) == (:Z, :T)
     @test Tables.columnnames(CausalTables.parents(coltbl2, :X)) == (:Z, :T)
     @test Tables.columnnames(CausalTables.parents(coltbl2, :Z)) == ()
-    @test Tables.columnnames(CausalTables.responseparents(coltbl2)) == (:X, :Z, :S, :T)
+    @test Tables.columnnames(CausalTables.responseparents(coltbl2)) == (:Z, :X, :T, :S)
     @test Tables.columnnames(CausalTables.parents(coltbl2, :Y)) == Tables.columnnames(CausalTables.responseparents(coltbl2))
 
     # Other convenience
@@ -96,6 +95,40 @@ end
 
     # Errors
[…]
