+++ src/causal_table.jl
@@ -21,9 +21,10 @@ function _process_causal_variable_names(treatment, response, causes)
         # Check that `causes` includes all treatment and response variables
         !all(map(x -> x ∈ keys(causes), vcat(treatment, response))) && throw(ArgumentError("`causes` must contain, at a minimum, a key for each Symbol contained in `treatment` or `response`. If a given treatment has no causes, set it equal to an empty Vector (i.e. `A = []`). If this error arose from using the `replace` function to change the `treatment` or `response` of a `CausalTable`, be sure to also update the `causes` attribute to reflect the new changes."))
 
-        # Ensure response does not cause any variables in the DAG
-        response_is_a_cause = any.(map(r -> r .∈ values(causes), response))
-        any(response_is_a_cause) && throw(ArgumentError("`causes` denotes a response variables causing a treatment variable, which is not allowed."))
+        # Ensure no responses cause any treatments in the DAG
+        treatment_causes = map(t -> causes[t], treatment)
+        response_is_a_cause = any.(map(r -> r .∈ treatment_causes, response))
+        any(response_is_a_cause) && throw(ArgumentError("`causes` denotes one or more response variables causing a treatment variable, which is not allowed."))
     end
 
     # Return fully processed causal variable names
@@ -362,32 +363,70 @@ function responseparents(o::CausalTable; collapse_parents = true)
     end
 end
 
+### Matrix utilities for the following
+
+# Iterator over matrix of CausalTables to turn them into matrices
+# If a CausalTable has empty data, return an empty matrix
+matrix(tbl::Matrix{CausalTable}) = map(x -> length(x.data) == 0 ? [;] : Tables.matrix(x), tbl)
+matrix(tbl::CausalTable) = Tables.matrix(tbl)
+
+function select_over_matrix(o::CausalTable, varnames; collapse_parents = true) 
+    # When possible, only select a single `CausalTable` representing the selection of all variables
+    if(collapse_parents && (length(o.response) + length(o.treatment) == 1 || all(x -> x == varnames[1,1], vec(varnames))))
+        return(select(o, varnames[1,1]))
+    # Otherwise, return a matrix of `CausalTables` representing the selection of each treatment-response pair
+    else
+        return([select(o, cn) for cn in varnames])
+    end
+end
+
+### Confounders ###
+
+"""
+    confoundernames(o::CausalTable)
+
+Outputs the confounder names of each response-treatment pair from the given `CausalTable` object.
+
+# Arguments
+- `o::CausalTable`: The `CausalTable` object from which to extract the confounder names of each treatment-response pair.
+
+# Returns
+A matrix of Vectors containing the confounder names of each treatment-response pair.
+"""
+function confoundernames(o::CausalTable)
+    # Extract the causes of each treatment variable as vector of vectors
+    treatment_parent_names = [o.causes[k] for k in keys(o.causes) if k ∈ o.treatment]
+    response_parent_names = [o.causes[k] for k in keys(o.causes) if k ∈ o.response]
+    return(hcat([[intersect(tpn, rpn) for tpn in treatment_parent_names] for rpn in response_parent_names]...))
+end
+
+"""
+    confoundernames(o::CausalTable, x::Symbol, y::Symbol)
+
+Outputs the names of the confounders of the causal relationship between `x` and `y` from the given `CausalTable` object.
+
+# Arguments
+- `o::CausalTable`: The `CausalTable` object from which to extract the confounder names.
+- `x::Symbol`, `y::Symbol`: The two variables whose confounders should be selected.
+
+# Returns
+A Vector of Symbols containing the names of the confounders between x and y.
+"""
[…]
