### 2025-01-10 130cbde
changes matrices to dictionaries
