+++ src/causal_table.jl
@@ -165,6 +165,8 @@ end
 DataAPI.nrow(o::CausalTable) = DataAPI.nrow(o.data)
 DataAPI.ncol(o::CausalTable) = DataAPI.ncol(o.data)
 
+Base.isempty(o::CausalTable) = Base.isempty(o.data)
+
 ### Causal-specific Features ###
 
 """
@@ -367,38 +369,27 @@ end
 
 # Iterator over matrix of CausalTables to turn them into matrices
 # If a CausalTable has empty data, return an empty matrix
-matrix(tbl::Matrix{CausalTable}) = map(x -> length(x.data) == 0 ? [;] : Tables.matrix(x), tbl)
-matrix(tbl::CausalTable) = Tables.matrix(tbl)
 
-function select_over_matrix(o::CausalTable, varnames; collapse_parents = true) 
+map_treatment_response_pairs(o::CausalTable, f::Function) = Dict(treat => Dict(resp => f(o, treat, resp) for resp in o.response) for treat in o.treatment)
+map_over_dicts(tbl, f::Function) = Dict(k => Dict(k2 => f(v2) for (k2, v2) in v) for (k, v) in tbl)
+
+function select_over_dicts(o::CausalTable, varnames; collapse_parents = true) 
     # When possible, only select a single `CausalTable` representing the selection of all variables
-    if(collapse_parents && (length(o.response) + length(o.treatment) == 1 || all(x -> x == varnames[1,1], vec(varnames))))
-        return(select(o, varnames[1,1]))
+    if(collapse_parents && 
+        (length(o.response) + length(o.treatment) == 1 || #  if there is only one treatment-response pair
+        all(x -> x == varnames[o.treatment[1]][o.response[1]], reduce(vcat, values.(values(varnames)))))) # if all pairs share the same set of variables
+
+        return(select(o, varnames[o.treatment[1]][o.response[1]]))
     # Otherwise, return a matrix of `CausalTables` representing the selection of each treatment-response pair
     else
-        return([select(o, cn) for cn in varnames])
+        return(map_over_dicts(varnames, x -> select(o, x)))
     end
 end
 
+matrix(tbl::Dict{Symbol, Dict{Symbol, CausalTable}}) = map_over_dicts(tbl, x -> isempty(x) ? [;] : Tables.matrix(x))
+matrix(tbl::CausalTable) = Tables.matrix(tbl)
+
 ### Confounders ###
-
-"""
-    confoundernames(o::CausalTable)
-
-Outputs the confounder names of each response-treatment pair from the given `CausalTable` object.
-
-# Arguments
-- `o::CausalTable`: The `CausalTable` object from which to extract the confounder names of each treatment-response pair.
-
-# Returns
-A matrix of Vectors containing the confounder names of each treatment-response pair.
-"""
-function confoundernames(o::CausalTable)
-    # Extract the causes of each treatment variable as vector of vectors
-    treatment_parent_names = [o.causes[k] for k in keys(o.causes) if k ∈ o.treatment]
-    response_parent_names = [o.causes[k] for k in keys(o.causes) if k ∈ o.response]
-    return(hcat([[intersect(tpn, rpn) for tpn in treatment_parent_names] for rpn in response_parent_names]...))
-end
 
 """
     confoundernames(o::CausalTable, x::Symbol, y::Symbol)
@@ -415,6 +406,19 @@ A Vector of Symbols containing the names of the confounders between x and y.
 confoundernames(o::CausalTable, x::Symbol, y::Symbol) = intersect(o.causes[x], o.causes[y])
 
 """
+    confoundernames(o::CausalTable)
+
+Outputs the confounder names of each response-treatment pair from the given `CausalTable` object.
+
+# Arguments
[…]
