### 2025-01-10 #7580 @olivierlabayle
> * involves

@salbalkus thank you for the reply and consideration!

Indeed, the present problem in Julia is the absence of a reference/consensus package for structural causal models (SCM). The present CausalTable structure will be insufficient to address most problems beyond the estimands implemented at the moment in this package. A non-specialist user, will want to quickly be able to understand if this package will be able to cover their use case. I do not believe it would be a good idea to have a dedicated structure for each currently known causal identification method (mediation, instrumental variables...) and the DAG with identification routines is probably a better idea.

At the moment, it seems to me the strength of the package lies in its simulation part for a specific subset of estimands rather than the CausalTable structure itself (could the name of the package be something like "CausalSimulations"?). Due to the aforementioned limitations, I feel the paper should be modified to reflect this. A few non exhaustive examples:

- line 29: "CausalTables.jl provides an interface to solve these two problems". I don't think it solves the first structural problem.
- line 44 "They do not provide a general causal-specific data processing framework like": what do you mean here by data processing? Maybe I missed something.
- line 93 "Users can simulate data from **any** SCM": this is incorrect as detailed above.
- ...

PS: If you would like to go further and integrate with TMLE.jl, it would probably be beneficial to create a core Julia package for SCMs. That way both TMLE.jl and this causal simulation package could depend on it.
