+++ paper/paper.qmd
@@ -16,7 +16,7 @@ authors:
 affiliations:
  - name: Department of Biostatistics, Harvard T.H. Chan School of Public Health
    index: 1
-date: 08 January 2025
+date: 20 January 2025
 bibliography: paper.bib
 format: markdown
 engine: julia
@@ -24,15 +24,15 @@ engine: julia
 
 # Summary
 
-Estimating the strength of causal relationships between variables is an important problem across many scientific disciplines. `CausalTables.jl` provides tools to (1) easily store and process tabular data endowed with causal structure, (2) simulate data from causal models for experimental testing, and (3) compute ground-truth estimates of causal parameters. Together, these functionalities expand the Julia ecosystem by supporting the development and experimental assessment of new statistical causal inference methods in Julia. 
+Estimating the strength of causal relationships between treatment and response variables is an important problem across many scientific disciplines. `CausalTables.jl` is a Julia package that helps statisticians and applied scientists create, manipulate, and simulate datasets labeled with relevant causal structure. Together, its functionalities expand the Julia ecosystem by supporting the development and experimental assessment of new statistical causal inference methods in Julia.
 
 # Statement of need
 
 The quantitative science of causal inference has emerged over the past three decades as a set of formalisms for studying cause-and-effect relationships between variables from observed data [@pearl2009causality; @hernan2020causal]. Causal inference techniques have helped applied scientists and decision-makers better understand important phenomena in fields ranging from health and medicine to politics and economics. New software tools for causal inference are being developed at a rapid pace, but in the Julia language, there currently do not exist auxiliary tools designed to support their development. `CausalTables.jl` aims to provide such a tool. 
 
-Attempts to implement and test causal inference methods in Julia face two major challenges. First, statistical causal inference requires data to be transformed in various ways based on the underlying causal structure, in order to be provided as input to other packages. For instance, causal methods using `MLJ.jl` [@blaom2020mlj] would necessitate extracting the variables hypothesized to cause the treatment and response as Tables; meanwhile; using `GLM.jl` [@bates2023glm] would require the same but as Matrix or `Formula` objects. Second, testing the performance of new estimators often requires simulating data for numerical experiments from a Structural Causal Model (SCM) [@pearl2009causality] so as to compare them to an underlying ground truth (encoded via interventions on the SCM). 
+Attempts to implement and test causal inference methods in Julia face two major challenges. First, statistical causal inference often requires extracting features from data based on their relationships with "treatment" and "response" variables; these might include confounders, mediators, or instruments. The format of these variables might even differ depending on downstream analysis package; for instance, `MLJ.jl` [@blaom2020mlj] requires input to be a Table, but `GLM.jl` [@bates2023glm] necessitates a Matrix or `Formula` object. Second, testing the performance of new estimators often requires simulating data for numerical experiments from a Structural Causal Model (SCM) [@pearl2009causality] so as to compare them to an underlying ground truth (encoded via interventions on the SCM). 
 
[…]
