### 2025-01-25 2febd5a
add additional DataGeneratingProcess constructors and merge function
