### 2025-01-27 215c902
test dgp reducing over O values
