### 2025-01-28 98cd27c
add empty rand method and update docs
