+++ docs/src/man/generating-data.md
@@ -4,7 +4,7 @@ When evaluating a causal inference method, we often want to test it on data from
 
 ## Defining a DataGeneratingProcess
 
-A data generating process describes a mechanism by which draws from random variables are simulated. It typically takes the form of a sequence of conditional distributions. CausalTables allows us to define a DGP as a DataGeneratingProcess object, which takes three arguments: the `names` of variables generated at each step, the `types` of these variables, and `funcs`, an array of functions of the form `(; O...) -> *some code`. 
+A data generating process describes a mechanism by which draws from random variables are simulated. It typically takes the form of a sequence of conditional distributions. CausalTables allows us to define a DGP as a DataGeneratingProcess object, which takes three arguments: the `names` of variables generated at each step, the `types` of these variables, and `funcs`, an array of functions of the form `O -> *some code`. 
 
 Suppose, for example, that we wanted to simulate data from the following DGP:
 
@@ -24,7 +24,6 @@ using CausalTables
 
 DataGeneratingProcess(
     [:W, :X, :Y],
-    [:distribution, :distribution, :distribution],
     [
         O -> DiscreteUniform(1, 5), 
         O -> (@. Normal(O.W, 1)),
@@ -43,7 +42,7 @@ However, a much more convenient way to define this DGP is using the `@dgp` macro
 using CausalTables
 distributions = @dgp(
         W ~ DiscreteUniform(1, 5),
-        X ~ (@. Normal(W, 1)),
+        X ~ Normal.(W, 1)),
         Y ~ (@. Normal(X + 0.2 * W, 1))
     )
 
@@ -51,15 +50,53 @@ distributions = @dgp(
 DataGeneratingProcess
 ```
 
-Note that with the `@dgp` macro, any symbol (that is, any string of characters prefixed by a colon, as in `:W` or `:X`) is automatically replaced with the corresponding previously-defined variable in the process. For instance, in `Normal(:W, 1)`, the `:W` will be replaced automatically with the distribution we defined as `W` earlier in the sequence. 
+Note that when using the `@dgp` macro, any symbol defined on the left side of an equation in the sequence can be used to pass in the output of a previous step on the right side. For example, in the above code, the symbol `W` is used to pass in the output of the first step to the second step. This works by metaprogramming which replaces `W` with `O.W` when the function is constructed by `@dgp`. 
+
+In this way, we can define virtually any DGP that can be expressed as a sequence of conditional distributions. For ease of use, one can still use the `O` object in the `@dgp` macro to pass in the output of all previous steps, which is especially useful for programmatically-defined DGPs. For example, the following code is equivalent to the above code:
+
+```jldoctest generation; output = false, filter = r"(?<=.{21}).*"s
+distributions = @dgp(
+        W ~ DiscreteUniform(1, 5),
+        X ~ Normal.(O[1], 1),
+        Y ~ Normal.(hcat(values(O)...) * [1, 0.2], 1)
+    )
+
+# output
+DataGeneratingProcess
+```
+
+In the first step, previous variables are accessed by index using `O[1]`, and in the third step, all previous variables are combined into a matrix by `hcat(values(O)...)`. Be careful when using these constructions, however, as they can make the code harder to read and understand. In some cases, it may be better to construct a `DataGeneratingProcess` manually using the constructor, for which several additional utilities are available. 
+
+For instance, if one wanted to generate a large number of variables with the same distribution, one could use the `DataGeneratingProcess` constructor without specifying variable names, in which case names will be automatically generated:
+
+```jldoctest generation; output = false, filter = r"(?<=.{75}).*"s
+
+many_distributions = DataGeneratingProcess(
+    [O -> Normal(0, 1) for _ in 1:100]
+)
+
+# output
[…]
