### 2025-01-28 #7580 @juliohm
Thank you for the reminder @lrnv. I believe that @olivierlabayle is much more involved with this specific ecosystem around causal models, and from my point of view after reading his review comments, it seems that the contribution of the package is not clear.

I would like to point out that not all packages need a JOSS entry. If @olivierlabayle believes this package fills an existing gap in the software literature, I can continue helping with the review. If, otherwise, we don't reach a consensus about novelty or contribution, I suggest we close the submission and give @salbalkus a chance to consider a future submission with a clearer problem statement and proposed solution.
