### 2025-01-28 #7580 @salbalkus
Hi everyone,

Apologies for the delay -- yesterday I pushed the latest update to CausalTables.jl which modifies the underlying functionality of the package considerably in response to comments from @olivierlabayle. Here is a list of the changes

First, I've replaced the `confounders` attribute in the `CausalTable` constructor with a new `causes::NamedTuple` attribute. The `causes` attribute is an edge list representing a (possibly partial) DAG; it includes, a minimum, labels of the causes of each attribute in `treatment` and `response`, *including "indirect" causes through other variables, not counting treatment and response.* This key principle is sufficient to construct most common identification sets, including confounders, mediators, and instruments, and also supports multiple treatment-response relationships with different sets of confounders. However, it is just as easy to extend the `CausalTable` package for novel types of problems that rely on more variables than just `treatment` and `response` to identify, as users can include in `causes` the causes of any variable contained in the data.

Because a `CausalTable` now contains a DAG structure supporting identification routines, I believe this should be sufficient to solve the "structural problem" @olivierlabayle. Building on this, in addition to `confounders`, I have implemented the `mediators` and `instruments` methods, which subset the columns of the CausalTable to include only their corresponding variables. These perform identification routines based on the partial DAG -- they don't require a new data structure.

This also addresses the comment about Line 44 -- one of the goals of this package is to make it easy to *extract* subsets of causally-relevant variables from a set of data with a single line of code, i.e. `confounders(ct)`. This is useful when developing new packages for causal inference

Additionally, I want to emphasize that this package actually *can* simulate any arbitrary SCMs, not just confounder-treatment-response structures, as long as they can be expressed as a sequence of conditional random draws (as we mention in the paper). For example, here's one where each outcome depends on neighboring treatments in a random graph

```
using Graphs
dgp = @dgp(
	n = 100,
	G = erdos_renyi(n, 0.5),
	X ~ MvNormal(n, 1),
	Y ~ MvNormal(Graphs.adjacency_matrix(G) * X, 1)
)
scm = StructuralCausalModel(dgp, :X, :Y)
rand(scm)
```

Note there are a few slight differences from the SCMs in the paper. In this setting, internal steps are "data adaptive" so the sample size is defined internally and `rand` can be called without specifying the sample size, since it is not meaningful. Also, `condensity` is now implicitly conditioning on the randomness not expressed by Distributions.jl objects, and we can obviously not extract the conditional density of these objects either. Technically, the constructor requires a `treatment` and `confounder` but these are really just window dressing; internally, they are only attached to the `CausalTable` output as labels. Happy to take any feedback on this.

As I mentioned, my goal is to build something that can serve as a reference package for causal models in Julia. Overall, I believe these changes should make the package flexible enough to handle the vast majority of problems in causal inference. Outside the scope of causal discovery (for which CausalInference.jl already exists), it's not really possible to have a causal inference problem that analyzes data but doesn't have at least one treatment and response, which is the only requirement for a `CausalTable` and technically isn't even needed for the `DataGeneratingProcess` to be defined. As it stands, new packages that build on CausalTables.jl should only need to take in a `CausalTable` as input in terms of the data needed to estimate a given causal estimand. I believe this is a novel contribution considering that no other package in Julia allows you to automatically extract relevant variables for identification directly from a dataset, nor extract conditional distributions from an SCM for ground-truth approximation.

PS: If `ct` is a `CausalTable`, one can extract a TMLE.jl-style SCM using something like the following:

```
SCM([k => v for (k, v) in pairs(ct.causes)])
```

and then pass in the CausalTable directly as the data. If it's all right with you, at some point I may open a pull request that adds functions in TMLE.jl to either allow users to write something like `ATE(ct)` to automatically construct the estimand from the information contained in a `CausalTable`, or `TMLE(ATE, ct)` to automatically construct the ATE object and pass it into the TMLE function.
