### 2025-01-28 3ace543
fix bug where replace overrides function in Base
