+++ docs/make.jl
@@ -2,17 +2,19 @@ using Documenter, CausalTables
 
 makedocs(
     sitename="CausalTables.jl",
-    pages = [
+    pages=[
         "Home" => "index.md",
-        "Turning data into a `CausalTable`" =>              "man/formatting.md",
-        "Generating data for statistical experiments" =>    "man/generating-data.md",
-        "Approximating ground truth causal estimands" =>    "man/estimands.md",
-        "Computing ground truth conditional distributions" =>    "man/ground-truth.md",
-        "Network summaries" =>                 "man/network-summaries.md",
-        "Full API" =>                                "man/api.md",
+        "Turning data into a `CausalTable`" => "man/formatting.md",
+        "Generating data for statistical experiments" => "man/generating-data.md",
+        "Approximating ground truth causal estimands" => "man/estimands.md",
+        "Computing ground truth conditional distributions" => "man/ground-truth.md",
+        "Network summaries" => "man/network-summaries.md",
+        "Full API" => "man/api.md",
     ]
 )
 
+
+
 deploydocs(
-    repo = "github.com/salbalkus/CausalTables.jl.git"
+    repo="github.com/salbalkus/CausalTables.jl.git"
 )
