### 2025-02-03 #7580 @olivierlabayle
Thank you @salbalkus and apologies for the delay, it is nice to see that CausalTable.jl now has a more flexible SCM interface and a forthcoming integration with TMLE.jl. In order to finish my review I will need a final information on the simulation part. I would be interested to know how the @dgp macro compares in flexibility to the Turing.jl @model or the @gen from Gen.jl.

For instance, unless I am missing something, I believe this works?

```julia
using Turing
@model function gdemo(n=100)
    G = erdos_renyi(n, 0.5)
    X ~ MvNormal(n, 1)
    Y ~ MvNormal(Graphs.adjacency_matrix(G) * X, 1)
end

model = gdemo()
intervened_1 = fix(model, X=ones(100));
intervened_0 = fix(model, X=zeros(100));
rand(model)
rand(intervened_1)
rand(intervened_0)
```

Given the maturity of these packages and the large overlap, could you make it precise how CausalTables compares to this state of the field? In particular:

- Can any Gen.jl or Turing.jl model also be defined within CausalTables.jl?
- As a user, why should I choose CausalTables over these packages to simulate data?
