### 2025-02-03 #7580 @salbalkus
@olivierlabayle Excellent questions!

> * Can any Gen.jl or Turing.jl model also be defined within CausalTables.jl?

In theory yes, though the syntax is somewhat more limited, as each line must take the form of a variable or distribution assignment. For example, in their tutorial `Gen.jl` defines

```
@gen function my_model(xs::Vector{Float64})
    slope ~ normal(0, 2)
    intercept ~ normal(0, 10)
    for (i, x) in enumerate(xs)
        {"y-$i"} ~ normal(slope * x + intercept, 1)
    end
end
```
In CausalTables.jl one cannot define a for loop within the `dgp` macro, but you could define a helper function calling the for-loop, perform a vector comprehension, or use dot notation. For example, in CausalTables the above could be written something like

```
@dgp(
    slope ~ Normal(0,2),
    intercept ~ Normal(0,1),
    y ~ (@. Normal(slope * x + intercept, 1))
)
```

Also, CausalTables.jl is not a fully-fledged PPL, since it doesn't implement any functionality to perform Bayesian inference, which is the primary purpose of Turing.jl and Gen.jl.

> * As a user, why should I choose CausalTables over these packages to simulate data?

The reason that I developed new simulation functionality instead of using Turing.jl or Gen.jl is to allow users to extract closed-form conditional distributions (as Distributions.jl objects) from any given step in the SCM, which doesn't seem to be possible with other packages (at least, I could not figure out how to do it). Many causal inference estimators involve conditional densities and/or conditional means -- if one were, for example, testing the statistical performance of a TMLE, they might want access to the true ("oracle") conditional density/propensity score and conditional mean for each observation in a given dataset, to make sure that their estimator works as intended with the oracles.

CausalTables makes it easy to extract the true conditional density and functionals of it using functions like `condensity`, `conmean`, or `propensity` automatically -- no need for the analyst write custom functions to compute the oracle for their data generating process! Hence, while there's no real difference in the random sampling of data, Turing.jl and Gen.jl don't allow closed-form distributions of each random variable (conditional on the randomly generated data that came before) to be extracted *after* data has been sampled, which is the strength of CausalTables.jl.

I'll update the paper and documentation soon to clarify the difference between this package and other PPL packages in Julia to ensure readers know the answers to these questions.
