### 2025-02-04 #7580 @olivierlabayle
Thank you for the clarification @salbalkus, as you say I think it is important to explain this distinctive feature. Ping me when the new version is ready and I'll finalise my review.
