+++ paper/paper.md
@@ -12,7 +12,7 @@ authors:
   name: Nima S. Hejazi
   orcid: 0000-0002-7127-2789
 bibliography: paper.bib
-date: 20 January 2025
+date: 6 February 2025
 tags:
 - Julia
 - statistics
@@ -27,12 +27,18 @@ toc-title: Table of contents
 
 Estimating the strength of causal relationships between treatment and
 response variables is an important problem across many scientific
-disciplines. `CausalTables.jl` is a Julia package that helps
-statisticians and applied scientists create, manipulate, and simulate
-datasets labeled with relevant causal structure. Together, its
-functionalities expand the Julia ecosystem by supporting the development
-and experimental assessment of new statistical causal inference methods
-in Julia.
+disciplines. `CausalTables.jl` is a Julia package providing two
+important functionalities to support causal inference in Julia. First,
+it provides the `CausalTable`, a Tables.jl-compliant data structure that
+bundles tabular data with causal structure. This allows users to
+automatically select variables, like treatment and confounders, that are
+often necessary as input to estimate causal quantities. Second, the
+package's `StructuralCausalModel` interface simplifies simulating data
+with a given causal structure -- and unlike existing simulation tools,
+users can extract the ground truth distributions and their properties
+conditional on the data generated in previous steps. In this way,
+CausalTables.jl makes it easier to develop and experimentally evaluate
+new statistical causal inference methods in Julia.
 
 # Statement of need
 
@@ -48,31 +54,38 @@ currently do not exist auxiliary tools designed to support their
 development. `CausalTables.jl` aims to provide such a tool.
 
 Attempts to implement and test causal inference methods in Julia face
-two major challenges. First, statistical causal inference often requires
-extracting features from data based on their relationships with
-"treatment" and "response" variables; these might include confounders,
-mediators, or instruments. The format of these variables might even
+two major challenges. First, causal estimation often requires
+identifying and extracting features from data based on their
+relationships with treatment and response variables; these might include
+confounders, mediators, or instruments. The format of these might even
 differ depending on downstream analysis package; for instance, `MLJ.jl`
 [@blaom2020mlj] requires input to be a Table, but `GLM.jl`
-[@bates2023glm] necessitates a Matrix or `Formula` object. Second,
-testing the performance of new estimators often requires simulating data
-for numerical experiments from a Structural Causal Model (SCM)
-[@pearl2009causality] so as to compare them to an underlying ground
-truth (encoded via interventions on the SCM).
+[@bates2023glm] necessitates a Matrix or `Formula` object. Second, since
+causal estimators are often tested on simulated data from a Structural
+Causal Model (SCM) [@pearl2009causality], it is often necessary to have
+access to the true ("oracle") conditional distributions of relevant
+variables in the SCM, as well as ground truth values of various causal
+estimands, in order to test whether a method works as intended.
 
-`CausalTables.jl` provides an interface to solve these two problems,
-simplifying the development of packages for statistical causal inference
-on tabular data in Julia. It implements a `CausalTable` interface that
-extends `Tables.jl` [@quinn2024tables] to also store necessary causal
-relationships between variables. In addition, the package implements a
-`StructuralCausalModel` interface for sampling from any SCM and
-computing ground-truth estimates of causal parameters. This interface
[…]
