+++ paper/paper.bib
@@ -187,4 +187,47 @@
   year={2020}
 }
 
+@inproceedings{turing,
+  author    = {Hong Ge and
+               Kai Xu and
+               Zoubin Ghahramani},
+  title     = {Turing: a language for flexible probabilistic inference},
+  booktitle = {International Conference on Artificial Intelligence and Statistics,
+               {AISTATS} 2018, 9-11 April 2018, Playa Blanca, Lanzarote, Canary Islands,
+               Spain},
+  pages     = {1682--1690},
+  year      = {2018},
+  url       = {http://proceedings.mlr.press/v84/ge18b.html},
+  biburl    = {https://dblp.org/rec/bib/conf/aistats/GeXG18},
+}
+
+@inproceedings{gen,
+ author = {Cusumano-Towner, Marco F. and Saad, Feras A. and Lew, Alexander K. and Mansinghka, Vikash K.},
+ title = {Gen: A General-purpose Probabilistic Programming System with Programmable Inference},
+ booktitle = {Proceedings of the 40th ACM SIGPLAN Conference on Programming Language Design and Implementation},
+ series = {PLDI 2019},
+ year = {2019},
+ isbn = {978-1-4503-6712-7},
+ location = {Phoenix, AZ, USA},
+ pages = {221--236},
+ numpages = {16},
+ url = {http://doi.acm.org/10.1145/3314221.3314642},
+ doi = {10.1145/3314221.3314642},
+ acmid = {3314642},
+ publisher = {ACM},
+ address = {New York, NY, USA},
+ keywords = {Markov chain Monte Carlo, Probabilistic programming, sequential Monte Carlo, variational inference},
+} 
+
+@article{Balkus2024,
+  doi = {10.48550/ARXIV.2412.02105},
+  url = {https://arxiv.org/abs/2412.02105},
+  author = {Balkus,  Salvador V. and Delaney,  Scott W. and Hejazi,  Nima S.},
+  keywords = {Methodology (stat.ME),  FOS: Computer and information sciences,  FOS: Computer and information sciences},
+  title = {The causal effects of modified treatment policies under network interference},
+  publisher = {arXiv},
+  year = {2024},
+  copyright = {arXiv.org perpetual,  non-exclusive license}
+}
+
 
