+++ paper/paper.md
@@ -29,16 +29,16 @@ Estimating the strength of causal relationships between treatment and
 response variables is an important problem across many scientific
 disciplines. `CausalTables.jl` is a Julia package providing two
 important functionalities to support causal inference in Julia. First,
-it provides the `CausalTable`, a Tables.jl-compliant data structure that
-bundles tabular data with causal structure. This allows users to
-automatically select variables, like treatment and confounders, that are
-often necessary as input to estimate causal quantities. Second, the
-package's `StructuralCausalModel` interface simplifies simulating data
-with a given causal structure -- and unlike existing simulation tools,
-users can extract the ground truth distributions and their properties
-conditional on the data generated in previous steps. In this way,
-CausalTables.jl makes it easier to develop and experimentally evaluate
-new statistical causal inference methods in Julia.
+it provides the `CausalTable`, which bundles tabular data with causal
+structure. This allows users to automatically subset and manipulate
+variables such as treatment and confounders that often must be provided
+as input to causal estimators. Second, the package's
+`StructuralCausalModel` interface simplifies running simulations with a
+given causal structure -- unlike existing simulation tools, users can
+extract the ground truth distributions and their properties conditional
+on the data generated in previous steps. In this way, CausalTables.jl
+makes it easier to develop and experimentally evaluate new statistical
+causal inference methods in Julia.
 
 # Statement of need
 
@@ -55,33 +55,33 @@ development. `CausalTables.jl` aims to provide such a tool.
 
 Attempts to implement and test causal inference methods in Julia face
 two major challenges. First, causal estimation often requires
-identifying and extracting features from data based on their
-relationships with treatment and response variables; these might include
-confounders, mediators, or instruments. The format of these might even
-differ depending on downstream analysis package; for instance, `MLJ.jl`
-[@blaom2020mlj] requires input to be a Table, but `GLM.jl`
-[@bates2023glm] necessitates a Matrix or `Formula` object. Second, since
-causal estimators are often tested on simulated data from a Structural
-Causal Model (SCM) [@pearl2009causality], it is often necessary to have
-access to the true ("oracle") conditional distributions of relevant
-variables in the SCM, as well as ground truth values of various causal
-estimands, in order to test whether a method works as intended.
+identifying features from data based on their relationships with
+treatment and response variables; these might include confounders,
+mediators, or instruments. Their required format might even differ
+depending on the downstream packages used; for instance, `MLJ.jl`
+[@blaom2020mlj] requires Table input, but `GLM.jl` [@bates2023glm]
+necessitates a Matrix or `Formula`. Second, since causal estimators are
+often tested on simulated data from a Structural Causal Model (SCM)
+[@pearl2009causality], one often desires access to the true ("oracle")
+conditional distributions of relevant variables in the SCM, as well as
+ground truth values of various causal estimands, in order to test
+whether a method works as intended.
 
-`CausalTables.jl` provides an interface to solve these two problems. The
-`CausalTable` interface extends `Tables.jl` [@quinn2024tables] to also
-store causal structure labels, including a type of directed acyclic
+`CausalTables.jl` provides a solution these two problems. The
+`CausalTable` interface extends `Tables.jl` [@quinn2024tables] to
[…]
