+++ paper/paper.bib
@@ -116,7 +116,7 @@
 }
 
 @misc{bates2023glm,
-  doi = {10.5281/ZENODO.3376013},
+  doi = {10.5281/zenodo.3376013},
   url = {https://zenodo.org/record/3376013},
   author = {Bates,  Douglas and Noack,  Andreas and Kornblith,  Simon and Bouchet-Valat,  Milan and Borregaard,  Michael Krabbe and Arslan,  Alex and White,  John Myles and Kleinschmidt,  Dave and Alday,  Phillip and Lynch,  Galen and Dunning,  Iain and Mogensen,  Patrick Kofod and Lendle,  Sam and {Dilum Aluthge} and {Mousum Dutta} and {Pdeffebach} and José Bayoán Santiago Calderón,  PhD and {Ayush Patnaik} and Born,  Benjamin and Setzler,  Bradley and DuBois,  Chris and Quinn,  Jacob and {Ondřej Slámečka} and Bastide,  Paul and {Viral B. Shah} and Anthony Blaom,  PhD and K\"{o}nig,  Bernhard},
   title = {JuliaStats/GLM.jl: v1.9.0},
@@ -140,7 +140,7 @@
 }
 
 @misc{Schauer2024,
-  doi = {10.5281/ZENODO.13684767},
+  doi = {10.5281/zenodo.13684767},
   author = {Schauer,  Moritz and Keller,  Martin and Wien\"{o}bst,  Marcel},
   title = {{CausalInference.jl}},
   publisher = {Zenodo},
@@ -220,12 +220,12 @@
 } 
 
 @article{Balkus2024,
-  doi = {10.48550/ARXIV.2412.02105},
+  doi = {10.48550/arxiv.2412.02105},
   url = {https://arxiv.org/abs/2412.02105},
   author = {Balkus,  Salvador V. and Delaney,  Scott W. and Hejazi,  Nima S.},
   keywords = {Methodology (stat.ME),  FOS: Computer and information sciences,  FOS: Computer and information sciences},
   title = {The causal effects of modified treatment policies under network interference},
-  publisher = {arXiv},
+  journal = {arXiv preprint},
   year = {2024},
   copyright = {arXiv.org perpetual,  non-exclusive license}
 }
