+++ paper/paper.qmd
@@ -24,15 +24,15 @@ engine: julia
 
 # Summary
 
-Estimating the strength of causal relationships between treatment and response variables is an important problem across many scientific disciplines. `CausalTables.jl` is a Julia package providing two important functionalities to support causal inference in Julia. First, it provides the `CausalTable`, which bundles tabular data with causal structure. This allows users to automatically subset and manipulate variables such as treatment and confounders that are often provided as input to causal estimators. Second, the package's `StructuralCausalModel` interface simplifies running simulations with a given causal structure -- unlike existing simulation tools, users can extract the ground truth distributions and their properties conditional on the data generated in previous steps. In this way, `CausalTables.jl` makes it easier to develop and experimentally evaluate new statistical causal inference methods in Julia.
+Estimating the strength of causal relationships between treatment and response variables is an important problem across many scientific disciplines. `CausalTables.jl` is a package providing two important functionalities to support causal inference in Julia. First, it provides the `CausalTable`, which bundles tabular data with causal structure. This allows users to automatically subset and manipulate variables such as treatment and confounders that are often provided as input to causal estimators. Second, the package's `StructuralCausalModel` interface simplifies running simulations with a given causal structure -- unlike existing simulation tools, users can extract the ground truth distributions and their properties conditional on the data generated in previous steps. In this way, `CausalTables.jl` makes it easier to develop and experimentally evaluate new statistical causal inference methods in Julia.
 
 # Statement of need
 
 The quantitative science of causal inference has emerged over the past three decades as a set of formalisms for studying cause-and-effect relationships between variables from observed data [@pearl2009causality; @hernan2020causal]. Causal inference techniques have helped scientists and decision-makers better understand important phenomena in fields ranging from medicine to economics. New software tools for causal inference are being developed at a rapid pace, but in the Julia language, there currently do not exist auxiliary tools designed to support their development. `CausalTables.jl` aims to provide such a tool. 
 
-Attempts to implement and test causal inference methods in Julia face two major challenges. First, causal estimation often requires identifying and modifying features from data based on their relationships with treatment and response variables; these might include confounders, mediators, or instruments. Their required format might even differ depending on the downstream packages used; for instance, `MLJ.jl` [@blaom2020mlj] requires Table input, but `GLM.jl` [@bates2023glm] necessitates a Matrix or `Formula`. Second, since causal estimators are often tested on simulated data from a Structural Causal Model (SCM) [@pearl2009causality], one often desires access to the true ("oracle") conditional distributions of relevant variables in the SCM, as well as ground truth values of various causal estimands, in order to test whether a method works correctly. 
+Implementing and testing causal inference methods in Julia involves two main challenges. First, causal estimation requires identifying and modifying features based on their relationships with treatment and response variables, which might include confounders, mediators, or instruments. Their required format may differ depending on downstream packages; for instance, `MLJ.jl` [@blaom2020mlj] requires Table input, while `GLM.jl` [@bates2023glm] needs a Matrix or `Formula`. Second, when evaluating a causal estimator on simulated data from a Structural Causal Model (SCM) [@pearl2009causality], one often desires access to the true ("oracle") conditional distributions of relevant variables in the SCM, as well as ground truth values of various causal estimands, in order to test whether the method works correctly. 
 
[…]
