### 2025-02-07 3edb070
make headers consistent in paper
