### 2025-02-07 049b379
make package formatting consistent in paper
