### 2025-02-07 #7580 @salbalkus
@olivierlabayle I've updated the paper, trying to provide a clearer description of the package's capabilities in the "Summary" section so users will know if it fits their use case. I've also fleshed out the description of the package's functionality and goals more clearly in the README and the home page of the documentation, including a paragraph telling users how it differs from similar packages. The documentation site will probably take a little bit to update -- once done, it should be ready for you to finalize your review.
