### 2025-02-07 #7580 @olivierlabayle
Thank you @salbalkus, I am happy with the state of the package, documentation and manuscript and have checked all my boxes!
