### 2025-02-12 #7580 @lrnv
@juliohm Indeed, now that @olivierlabayle reached a decision on the scholarly effort behind the contribution, would you like to take another look ?
