### 2025-02-12 #7580 @juliohm
Amazing work @salbalkus. The new documentation and paper are crystal clear, and the examples worked like a charm in my local machine.

Feel free to proceed to publication. I checked all boxes 💯 👏🏽
