+++ Project.toml
@@ -1,7 +1,7 @@
 name = "CausalTables"
 uuid = "6af48e0c-efc2-4bf7-a92f-a553ccf79fd6"
 authors = ["Salvador Balkus", "Nima Hejazi"]
-version = "1.2.1"
+version = "1.2.2"
 
 [deps]
 DataAPI = "9a962f9c-6df0-11e9-0e5d-c546b8b5ee8a"
