### 2025-02-13 #7580 @salbalkus
@lrnv I've created a new release of the software, version number 1.2.2. The Zenodo DOI is https://doi.org/10.5281/zenodo.14867116. The authors, affiliations, title, and license should match.
