+++ paper/paper.bib
@@ -81,7 +81,7 @@
                   Jan Drugowitsch and
                   Moritz Schauer and
                   other contributors},
-  title        = {{JuliaStats/Distributions.jl: a Julia package for probability distributions and associated functions}},
+  title        = {{JuliaStats/Distributions.jl: a {Julia} package for probability distributions and associated functions}},
   month        = jul,
   year         = 2019,
   doi          = {10.5281/zenodo.2647458},
@@ -106,7 +106,7 @@
 }
 
 @article{blaom2020mlj,
-  title={MLJ: A Julia package for composable machine learning},
+  title={MLJ: A {Julia} package for composable machine learning},
   author={Blaom, Anthony D and Kiraly, Franz and Lienart, Thibaut and Simillides, Yiannis and Arenas, Diego and Vollmer, Sebastian J},
   journal={Journal of Open Source Software},
   volume={5},
