### 2025-02-14 c50696e
update paper citations
