+++ paper/paper.qmd
@@ -16,7 +16,7 @@ authors:
 affiliations:
  - name: Department of Biostatistics, Harvard T.H. Chan School of Public Health
    index: 1
-date: 13 February 2025
+date: 14 February 2025
 bibliography: paper.bib
 format: markdown
 engine: julia
