+++ paper/paper.bib
@@ -1,15 +1,17 @@
-@misc{hernan2020causal,
-  title={Causal inference},
+@book{hernan2020causal,
+  title={{Causal Inference: What If}},
   author={Hern{\'a}n, Miguel A and Robins, James M},
   year={2020},
-  publisher={CRC Boca Raton, FL}
+  publisher={CRC Boca Raton, FL},
+  isbn={9781420076165}
 }
 
 @book{pearl2009causality,
   title={Causality},
   author={Pearl, Judea},
   year={2009},
-  publisher={Cambridge university press}
+  publisher={Cambridge University Press},
+  isbn={9780521895606}
 }
 
 @software{quinn2024tables,
@@ -106,9 +108,12 @@
 }
 
 @article{blaom2020mlj,
-  title={MLJ: A {Julia} package for composable machine learning},
+  title={{MLJ}: A {Julia} package for composable machine learning},
   author={Blaom, Anthony D and Kiraly, Franz and Lienart, Thibaut and Simillides, Yiannis and Arenas, Diego and Vollmer, Sebastian J},
   journal={Journal of Open Source Software},
+  url = {https://doi.org/10.21105/joss.02704},
+  doi = {10.21105/joss.02704},
+  publisher = {The Open Journal},
   volume={5},
   number={55},
   pages={2704},
@@ -177,13 +182,15 @@
   author       = {Chen, Huigang and Harinen, Totte and Lee, Jeong-Yoon and
                   Yung, Mike and Zhao, Zhenyu},
   year         =  {2020},
-  journal       = {arXiv preprint arXiv:2002.11631}
+  journal       = {arXiv preprint},
+  doi = {arXiv:2002.11631}
 }
 
 @article{dowhy,
-  title={DoWhy: An End-to-End Library for Causal Inference},
+  title={{DoWhy}: An End-to-End Library for Causal Inference},
   author={Sharma, Amit and Kiciman, Emre},
-  journal={arXiv preprint arXiv:2011.04216},
+  journal={arXiv preprint},
+  doi = {arXiv:2011.04216},
   year={2020}
 }
 
@@ -195,10 +202,11 @@
   booktitle = {International Conference on Artificial Intelligence and Statistics,
                {AISTATS} 2018, 9-11 April 2018, Playa Blanca, Lanzarote, Canary Islands,
                Spain},
+  publisher = {Proceedings of Machine Learning Research},
+  doi = {https://doi.org/10.17863/CAM.42246},
   pages     = {1682--1690},
   year      = {2018},
   url       = {http://proceedings.mlr.press/v84/ge18b.html},
-  biburl    = {https://dblp.org/rec/bib/conf/aistats/GeXG18},
 }
 
 @inproceedings{gen,
@@ -220,7 +228,7 @@
 } 
 
 @article{Balkus2024,
-  doi = {10.48550/arxiv.2412.02105},
+  doi = {arxiv.2412.02105},
   url = {https://arxiv.org/abs/2412.02105},
   author = {Balkus,  Salvador V. and Delaney,  Scott W. and Hejazi,  Nima S.},
   keywords = {Methodology (stat.ME),  FOS: Computer and information sciences,  FOS: Computer and information sciences},
