### 2025-02-14 bf3a9b3
fix paper.bib typos
