+++ paper/paper.bib
@@ -86,6 +86,7 @@
   title        = {{JuliaStats/Distributions.jl: a {Julia} package for probability distributions and associated functions}},
   month        = jul,
   year         = 2019,
+  publisher    = {Zenodo},
   doi          = {10.5281/zenodo.2647458},
   url          = {https://doi.org/10.5281/zenodo.2647458}
 }
@@ -203,7 +204,7 @@
                {AISTATS} 2018, 9-11 April 2018, Playa Blanca, Lanzarote, Canary Islands,
                Spain},
   publisher = {Proceedings of Machine Learning Research},
-  doi = {https://doi.org/10.17863/CAM.42246},
+  doi = {10.17863/CAM.42246},
   pages     = {1682--1690},
   year      = {2018},
   url       = {http://proceedings.mlr.press/v84/ge18b.html},
