+++ paper/paper.bib
@@ -184,14 +184,14 @@
                   Yung, Mike and Zhao, Zhenyu},
   year         =  {2020},
   journal       = {arXiv preprint},
-  doi = {arXiv:2002.11631}
+  doi = {10.48550/arXiv.2002.11631}
 }
 
 @article{dowhy,
   title={{DoWhy}: An End-to-End Library for Causal Inference},
   author={Sharma, Amit and Kiciman, Emre},
   journal={arXiv preprint},
-  doi = {arXiv:2011.04216},
+  doi = {10.48550/arXiv:2011.04216},
   year={2020}
 }
 
@@ -229,7 +229,7 @@
 } 
 
 @article{Balkus2024,
-  doi = {arxiv.2412.02105},
+  doi = {10.48550/arXiv.2412.02105},
   url = {https://arxiv.org/abs/2412.02105},
   author = {Balkus,  Salvador V. and Delaney,  Scott W. and Hejazi,  Nima S.},
   keywords = {Methodology (stat.ME),  FOS: Computer and information sciences,  FOS: Computer and information sciences},
