+++ paper/paper.bib
@@ -191,7 +191,7 @@
   title={{DoWhy}: An End-to-End Library for Causal Inference},
   author={Sharma, Amit and Kiciman, Emre},
   journal={arXiv preprint},
-  doi = {10.48550/arXiv:2011.04216},
+  doi = {10.48550/arXiv.2011.04216},
   year={2020}
 }
 
