### 2025-02-14 #7580 @salbalkus
Hi @lrnv I fixed the missing DOIs and some of the ones that appeared skipped. The remaining skipped DOI are all either books ("Causal Inference: What If", "Causality", and "Targeted Learning in R") or software packages without a DOI ("TMLE.jl", "CausalELM.jl", and "causaldag").
