### 2025-02-21 #7580 @crvernon
:wave: @salbalkus - I just need you to make the following change before I accept this for publication:

In the paper:

- [ ] LINE 148:  The "r" in "The r package..." should be capitalized.  You can do this by placing curly brackets around characters you wish to maintain the formatting of in your bib file.

Let me know when this has been fixed.  Thanks.
