### 2025-02-22 928a824
fix bibliography capitalization
