+++ paper/paper.bib
@@ -162,7 +162,7 @@
 }
 
  @article{Textor2017, 
-  title={Robust causal inference using directed acyclic graphs: the R package ‘dagitty’}, 
+  title={Robust causal inference using directed acyclic graphs: the {R} package ‘dagitty’}, 
   ISSN={1464-3685}, url={http://dx.doi.org/10.1093/ije/dyw341}, 
   DOI={10.1093/ije/dyw341}, journal={International Journal of Epidemiology}, 
   publisher={Oxford University Press (OUP)}, 
