### 2025-02-22 #7580 @salbalkus
Thanks for catching this @crvernon, I've capitalized the R so it should be all set now!
